refactoring

+++ 2DGameDevLib/GameData.h
@@ -22,10 +22,10 @@ class GameData : public gamelib::GameWorldData
 {
 public:
 	static GameData* Get();
-	GameData(const GameData&) = delete; // copy constructor 
-	GameData(const GameData&&) = delete; // move constructor
-	GameData& operator=(const GameData&) = delete; // assignment
-	GameData& operator=(const GameData&&) = delete; // move assignment	
+	GameData(const GameData&) = delete;
+	GameData(const GameData&&) = delete;
+	GameData& operator=(const GameData&) = delete;
+	GameData& operator=(const GameData&&) = delete;
 	~GameData() { instance = nullptr; }
 
 	void AddRoom(const std::shared_ptr<Room>& room);
@@ -36,9 +36,12 @@ public:
 	void RemoveEnemy(const std::shared_ptr<Enemy>& enemy);
 
 	std::shared_ptr<Room> GetRoomByIndex(int roomNumber);
-	[[nodiscard]] std::shared_ptr<Player> GetPlayer() const;
-	[[nodiscard]] unsigned int CountPickups() const { return pickups.size(); }
-	[[nodiscard]] bool IsGameWon() const { return isGameWon; }
+	[[nodiscard]]
+	std::shared_ptr<Player> GetPlayer() const;
+	[[nodiscard]]
+	unsigned int CountPickups() const { return static_cast<unsigned int>(pickups.size()); }
+	[[nodiscard]]
+	bool IsGameWon() const { return isGameWon; }
 	void SetGameWon(const bool yesNo) { isGameWon = yesNo; }
 	void AddGameObject(const std::shared_ptr<gamelib::GameObject>& gameObject);
 	void RemoveGameObject(const std::shared_ptr<gamelib::GameObject>& gameObject);
